ci: add E2E smoke test stage; remove redundant end_to_end_tests
Add a docker-compose-based E2E job to CI that builds the full stack,
waits for health, and runs the existing Rust E2E test suite. Known-issue
Prometheus API tests (range query, series, label selectors, aggregation)
are factored into a separate file with distinct ignore tags so they are
excluded from CI without relying on soft assertions.

- Add `e2e` job to ci.yml (runs after `test`, 30min timeout)
- Add dogfooding self-telemetry verification test
- Factor prometheus_known_issues_tests.rs out of prometheus_api_tests.rs
- Delete tests/end_to_end_tests.rs (all 7 tests redundant with dedicated files)
- Upload container logs as artifacts on all runs
