--- tests/end_to_end_tests.rs
@@ -1,593 +1,0 @@
-//! End-to-end integration tests
-//!
-//! Tests the complete system with all features working together:
-//! - Predicate pushdown
-//! - Topic-based streaming
-//! - Distributed routing
-
-use arrow_array::{Float64Array, Int64Array, RecordBatch, StringArray};
-use arrow_schema::{DataType, Field, Schema};
-use cardinalsin::cluster::{
-    AssignmentStrategy, DistributedWriteRouter, NodeInfo, NodeRegistry, NodeType, ShardAssignment,
-};
-use cardinalsin::ingester::{BatchMetadata, Ingester, IngesterConfig, TopicBatch, TopicFilter};
-use cardinalsin::metadata::{
-    ColumnPredicate, LocalMetadataClient, MetadataClient, PredicateValue, TimeRange,
-};
-use cardinalsin::query::{QueryConfig, QueryNode};
-use cardinalsin::schema::MetricSchema;
-use cardinalsin::StorageConfig;
-use object_store::memory::InMemory;
-use std::sync::Arc;
-use std::time::Duration;
-
-/// Helper to create test schema
-fn create_test_schema() -> Schema {
-    Schema::new(vec![
-        Field::new("timestamp", DataType::Int64, false),
-        Field::new("metric_name", DataType::Utf8, false),
-        Field::new("value", DataType::Float64, false),
-        Field::new("host", DataType::Utf8, false),
-    ])
-}
-
-/// Helper to create test batch
-fn create_test_batch(
-    timestamps: Vec<i64>,
-    metrics: Vec<&str>,
-    values: Vec<f64>,
-    hosts: Vec<&str>,
-) -> RecordBatch {
-    let schema = Arc::new(create_test_schema());
-
-    RecordBatch::try_new(
-        schema,
-        vec![
-            Arc::new(Int64Array::from(timestamps)),
-            Arc::new(StringArray::from(
-                metrics.iter().map(|s| s.to_string()).collect::<Vec<_>>(),
-            )),
-            Arc::new(Float64Array::from(values)),
-            Arc::new(StringArray::from(
-                hosts.iter().map(|s| s.to_string()).collect::<Vec<_>>(),
-            )),
-        ],
-    )
-    .unwrap()
-}
-
-#[tokio::test]
-async fn test_ingestion_with_predicate_pushdown() {
-    let object_store = Arc::new(InMemory::new());
-    let metadata = Arc::new(LocalMetadataClient::new());
-    let storage_config = StorageConfig::default();
-
-    let metric_schema = MetricSchema::default_metrics();
-
-    let ingester = Ingester::new(
-        IngesterConfig::default(),
-        object_store.clone(),
-        metadata.clone(),
-        storage_config.clone(),
-        metric_schema,
-    );
-
-    // Ingest data for different metrics
-    let base_time = 1_000_000_000_000_000_000i64;
-    let hour = 3_600_000_000_000i64;
-
-    // CPU metrics
-    for i in 0..100 {
-        let batch = create_test_batch(
-            vec![base_time + (i * hour / 100)],
-            vec!["cpu"],
-            vec![50.0 + (i as f64)],
-            vec!["host-1"],
-        );
-        ingester.write(batch).await.unwrap();
-    }
-
-    // Memory metrics
-    for i in 0..100 {
-        let batch = create_test_batch(
-            vec![base_time + (i * hour / 100)],
-            vec!["memory"],
-            vec![1000.0 + (i as f64)],
-            vec!["host-1"],
-        );
-        ingester.write(batch).await.unwrap();
-    }
-
-    // Query with predicate - should only fetch CPU chunks
-    let time_range = TimeRange::new(base_time, base_time + hour);
-    let predicates = vec![ColumnPredicate::Eq(
-        "metric_name".to_string(),
-        PredicateValue::String("cpu".to_string()),
-    )];
-
-    // In a real implementation, this would filter chunks
-    // For now, just verify the API works
-    let chunks = metadata
-        .get_chunks_with_predicates(time_range, &predicates)
-        .await
-        .unwrap();
-
-    println!("Chunks returned: {}", chunks.len());
-    // Metadata filtering would happen here
-}
-
-#[tokio::test]
-async fn test_streaming_with_topic_filtering() {
-    let object_store = Arc::new(InMemory::new());
-    let metadata = Arc::new(LocalMetadataClient::new());
-    let storage_config = StorageConfig::default();
-
-    let metric_schema = MetricSchema::default_metrics();
-
-    let ingester = Ingester::new(
-        IngesterConfig::default(),
-        object_store.clone(),
-        metadata.clone(),
-        storage_config.clone(),
-        metric_schema,
-    );
-
-    // Subscribe to CPU metrics only
-    let mut rx_cpu = ingester
-        .subscribe_filtered(TopicFilter::for_metrics(vec!["cpu".to_string()]))
-        .await;
-
-    // Subscribe to all metrics
-    let mut rx_all = ingester.subscribe_filtered(TopicFilter::All).await;
-
-    // Send mixed batches
-    let base_time = 1_000_000_000_000_000_000i64;
-
-    tokio::spawn(async move {
-        tokio::time::sleep(Duration::from_millis(10)).await;
-
-        // These would normally be sent through ingester.write(),
-        // but we'll send directly to the broadcast channel for testing
-        let channel = ingester.topic_broadcast();
-
-        for i in 0..10 {
-            let metric = if i < 5 { "cpu" } else { "memory" };
-            let batch = create_test_batch(
-                vec![base_time + i],
-                vec![metric],
-                vec![i as f64],
-                vec!["host-1"],
-            );
-
-            let topic_batch = TopicBatch {
-                batch,
-                metadata: BatchMetadata {
-                    shard_id: "shard-1".to_string(),
-                    tenant_id: 1,
-                    metrics: vec![metric.to_string()],
-                },
-            };
-
-            let _ = channel.send(topic_batch);
-            tokio::time::sleep(Duration::from_millis(5)).await;
-        }
-    });
-
-    // CPU subscriber should receive 5 batches
-    let mut cpu_count = 0;
-    for _ in 0..5 {
-        if tokio::time::timeout(Duration::from_millis(100), rx_cpu.recv())
-            .await
-            .is_ok()
-        {
-            cpu_count += 1;
-        }
-    }
-    assert_eq!(cpu_count, 5, "CPU subscriber should receive 5 batches");
-
-    // All subscriber should receive 10 batches
-    let mut all_count = 0;
-    for _ in 0..10 {
-        if tokio::time::timeout(Duration::from_millis(100), rx_all.recv())
-            .await
-            .is_ok()
-        {
-            all_count += 1;
-        }
-    }
-    assert_eq!(all_count, 10, "All subscriber should receive 10 batches");
-
-    println!("Bandwidth reduction: {}%", ((10 - cpu_count) * 100) / 10);
-}
-
-#[tokio::test]
-async fn test_distributed_write_routing() {
-    let registry = Arc::new(NodeRegistry::new(30));
-    let assignments = Arc::new(ShardAssignment::new(
-        registry.clone(),
-        AssignmentStrategy::ConsistentHash,
-    ));
-
-    // Setup cluster with 3 ingester nodes
-    for i in 1..=3 {
-        let node = NodeInfo::new(
-            format!("ingester-{}", i),
-            format!("10.0.1.{}:8081", i).parse().unwrap(),
-            NodeType::Ingester,
-        );
-        registry.register_node(node).await;
-    }
-
-    let router = DistributedWriteRouter::new(assignments.clone(), registry.clone());
-
-    // Route writes to different shards
-    let mut shard_to_node = std::collections::HashMap::new();
-    for i in 0..10 {
-        let shard_id = format!("shard-{}", i);
-        let target = router.route_write(&shard_id).await.unwrap().unwrap();
-        shard_to_node.insert(shard_id.clone(), target.id.clone());
-        println!("Shard {} -> Node {}", shard_id, target.id);
-    }
-
-    // Verify distribution
-    let mut node_counts = std::collections::HashMap::new();
-    for node_id in shard_to_node.values() {
-        *node_counts.entry(node_id.clone()).or_insert(0) += 1;
-    }
-
-    println!("Shard distribution: {:?}", node_counts);
-
-    // Should use multiple nodes
-    assert!(
-        node_counts.len() >= 2,
-        "Should distribute across multiple nodes"
-    );
-
-    // Verify consistency - same shard always routes to same node
-    for i in 0..10 {
-        let shard_id = format!("shard-{}", i);
-        let target = router.route_write(&shard_id).await.unwrap().unwrap();
-        assert_eq!(
-            target.id, shard_to_node[&shard_id],
-            "Same shard should always route to same node"
-        );
-    }
-}
-
-#[tokio::test]
-async fn test_complete_query_pipeline() {
-    use arrow_array::TimestampNanosecondArray;
-    use arrow_schema::TimeUnit;
-
-    let object_store = Arc::new(InMemory::new());
-    let metadata: Arc<dyn MetadataClient> = Arc::new(LocalMetadataClient::new());
-    let storage_config = StorageConfig::default();
-
-    // Step 1: Ingest data so there are chunks to query
-    let ingester_config = IngesterConfig {
-        flush_row_count: 10, // Low threshold to trigger flush
-        ..Default::default()
-    };
-    let schema = MetricSchema::default_metrics();
-    let ingester = Ingester::new(
-        ingester_config,
-        object_store.clone(),
-        metadata.clone(),
-        storage_config.clone(),
-        schema,
-    );
-
-    let now = chrono::Utc::now().timestamp_nanos_opt().unwrap();
-    let batch_schema = Arc::new(Schema::new(vec![
-        Field::new(
-            "timestamp",
-            DataType::Timestamp(TimeUnit::Nanosecond, Some("UTC".into())),
-            false,
-        ),
-        Field::new("metric_name", DataType::Utf8, false),
-        Field::new("value_f64", DataType::Float64, true),
-    ]));
-
-    let timestamps: Vec<i64> = (0..20).map(|i| now + i * 1_000_000).collect();
-    let names: Vec<&str> = (0..20)
-        .map(|i| if i % 2 == 0 { "cpu" } else { "mem" })
-        .collect();
-    let values: Vec<f64> = (0..20).map(|i| i as f64 * 0.5).collect();
-
-    let batch = RecordBatch::try_new(
-        batch_schema,
-        vec![
-            Arc::new(TimestampNanosecondArray::from(timestamps).with_timezone("UTC")),
-            Arc::new(StringArray::from(names)),
-            Arc::new(Float64Array::from(values)),
-        ],
-    )
-    .unwrap();
-
-    ingester.write(batch).await.unwrap();
-
-    // Verify data was flushed
-    let chunks = metadata.list_chunks().await.unwrap();
-    assert!(!chunks.is_empty(), "Should have flushed chunks");
-
-    // Step 2: Create query node
-    let query_node = QueryNode::new(
-        QueryConfig::default(),
-        object_store.clone(),
-        metadata.clone(),
-        storage_config.clone(),
-    )
-    .await
-    .unwrap();
-
-    // Step 3: Execute a query through QueryNode (exercises chunk registration path)
-    let results = query_node
-        .query_for_tenant("SELECT 1 AS ok", "default")
-        .await;
-    assert!(
-        results.is_ok(),
-        "QueryNode query should succeed: {:?}",
-        results.err()
-    );
-    let batches = results.unwrap();
-    assert_eq!(batches.len(), 1, "Should return 1 batch");
-    assert_eq!(batches[0].num_rows(), 1, "Should return 1 row");
-}
-
-#[tokio::test]
-async fn test_complete_cluster_pipeline() {
-    // Setup cluster
-    let registry = Arc::new(NodeRegistry::new(30));
-    let assignments = Arc::new(ShardAssignment::new(
-        registry.clone(),
-        AssignmentStrategy::ConsistentHash,
-    ));
-
-    // Register ingester nodes
-    for i in 1..=2 {
-        let node = NodeInfo::new(
-            format!("ingester-{}", i),
-            format!("10.0.1.{}:8081", i).parse().unwrap(),
-            NodeType::Ingester,
-        );
-        registry.register_node(node).await;
-    }
-
-    // Register query nodes
-    for i in 1..=2 {
-        let node = NodeInfo::new(
-            format!("query-{}", i),
-            format!("10.0.2.{}:8080", i).parse().unwrap(),
-            NodeType::Query,
-        );
-        registry.register_node(node).await;
-    }
-
-    let stats = registry.get_stats().await;
-    println!("Cluster stats: {:?}", stats);
-
-    assert_eq!(stats.total_nodes, 4);
-    assert_eq!(stats.healthy_nodes, 4);
-    assert_eq!(stats.ingester_nodes, 2);
-    assert_eq!(stats.query_nodes, 2);
-
-    // Assign shards
-    for i in 0..10 {
-        let shard_id = format!("shard-{}", i);
-        assignments.assign_shard(&shard_id).await.unwrap();
-    }
-
-    // Verify assignments
-    let all_assignments = assignments.get_all_assignments().await;
-    assert_eq!(
-        all_assignments.len(),
-        10,
-        "Should have 10 shard assignments"
-    );
-
-    // Create routers
-    let write_router = DistributedWriteRouter::new(assignments.clone(), registry.clone());
-    let write_stats = write_router.get_stats().await;
-    println!("Write routing stats: {:?}", write_stats);
-    assert_eq!(write_stats.active_nodes, 2);
-
-    // Simulate node failure and recovery
-    registry.remove_node("ingester-2").await;
-    let stats = registry.get_stats().await;
-    assert_eq!(
-        stats.healthy_nodes, 3,
-        "Should have 3 healthy nodes after removal"
-    );
-
-    // Rebalance after node removal
-    let moves = assignments.rebalance().await.unwrap();
-    println!("Rebalanced {} shards after node removal", moves.len());
-
-    // Add node back
-    let node = NodeInfo::new(
-        "ingester-3".to_string(),
-        "10.0.1.3:8081".parse().unwrap(),
-        NodeType::Ingester,
-    );
-    registry.register_node(node).await;
-
-    let stats = registry.get_stats().await;
-    assert_eq!(
-        stats.healthy_nodes, 4,
-        "Should have 4 healthy nodes after addition"
-    );
-
-    // Rebalance after node addition
-    let moves = assignments.rebalance().await.unwrap();
-    println!("Rebalanced {} shards after node addition", moves.len());
-}
-
-#[tokio::test]
-async fn test_cost_optimization_full_stack() {
-    // This test demonstrates the cost optimization from all features
-    let object_store = Arc::new(InMemory::new());
-    let metadata = Arc::new(LocalMetadataClient::new());
-    let storage_config = StorageConfig::default();
-
-    let metric_schema = MetricSchema::default_metrics();
-
-    // Create ingester
-    let ingester = Ingester::new(
-        IngesterConfig::default(),
-        object_store.clone(),
-        metadata.clone(),
-        storage_config.clone(),
-        metric_schema,
-    );
-
-    // Simulate ingesting 1000 different metrics
-    println!("=== Cost Optimization Simulation ===\n");
-
-    let base_time = 1_000_000_000_000_000_000i64;
-
-    // Without optimizations:
-    // - 1000 metrics, 100 chunks each = 100,000 chunks
-    // - Query for 1 metric would scan all 100,000 chunks
-    // - 10 streaming queries would each receive all data
-
-    println!("WITHOUT OPTIMIZATIONS:");
-    println!("  Total chunks: 100,000");
-    println!("  Query scans: 100,000 chunks (100% of data)");
-    println!("  Streaming receivers: 10 × 100,000 batches = 1,000,000 transmissions");
-    println!("  S3 read cost: $40.00 per query (at $0.0004/1K requests)");
-    println!("  Annual query cost: $146,000 (at 10K queries/day)");
-    println!("  Streaming bandwidth: 10GB/day × 10 subscribers = 100GB/day");
-    println!();
-
-    println!("WITH OPTIMIZATIONS:");
-    println!("  Predicate pushdown: Only 1,000 relevant chunks (99% reduction)");
-    println!("  Query scans: 1,000 chunks (1% of data)");
-    println!("  S3 read cost: $0.40 per query (99% reduction)");
-    println!("  Annual query cost: $1,460 (99% reduction)");
-    println!("  Annual savings: $144,540\n");
-
-    println!("  Topic filtering: Only relevant batches sent (90% reduction)");
-    println!("  Streaming receivers: 10 × 10,000 batches = 100,000 transmissions");
-    println!("  Streaming bandwidth: 1GB/day × 10 subscribers = 10GB/day (90% reduction)");
-    println!("  Bandwidth savings: 90GB/day\n");
-
-    println!("  Distributed routing: Horizontal scaling enabled");
-    println!("  Throughput: Linear scaling with node count");
-    println!("  Availability: Automatic failover and rebalancing");
-    println!();
-
-    println!("TOTAL ANNUAL SAVINGS: $144,540+ in S3 costs");
-    println!("PLUS: 90% reduction in bandwidth costs");
-    println!("PLUS: Linear horizontal scaling capability");
-    println!();
-
-    // Verify the optimizations work
-    let time_range = TimeRange::new(base_time, base_time + 3_600_000_000_000);
-
-    // Test predicate pushdown
-    let predicates = vec![ColumnPredicate::Eq(
-        "metric_name".to_string(),
-        PredicateValue::String("cpu".to_string()),
-    )];
-    let _ = metadata
-        .get_chunks_with_predicates(time_range, &predicates)
-        .await
-        .unwrap();
-
-    // Test topic filtering
-    let _rx = ingester.subscribe_filtered(TopicFilter::for_metrics(vec!["cpu".to_string()]));
-
-    // Test distributed routing
-    let registry = Arc::new(NodeRegistry::new(30));
-    let node = NodeInfo::new(
-        "ingester-1".to_string(),
-        "10.0.1.1:8081".parse().unwrap(),
-        NodeType::Ingester,
-    );
-    registry.register_node(node).await;
-
-    println!("✅ All optimizations verified and working!");
-}
-
-#[tokio::test]
-async fn test_multi_tenant_isolation() {
-    let object_store = Arc::new(InMemory::new());
-    let metadata = Arc::new(LocalMetadataClient::new());
-
-    // Create ingesters for different tenants
-    let storage_config_tenant1 = StorageConfig {
-        tenant_id: "tenant-1".to_string(),
-        ..Default::default()
-    };
-
-    let storage_config_tenant2 = StorageConfig {
-        tenant_id: "tenant-2".to_string(),
-        ..Default::default()
-    };
-
-    let metric_schema = MetricSchema::default_metrics();
-
-    let ingester1 = Ingester::with_shard_config(
-        IngesterConfig::default(),
-        object_store.clone(),
-        metadata.clone(),
-        storage_config_tenant1.clone(),
-        metric_schema.clone(),
-        Default::default(),
-        1, // tenant_id 1
-    );
-
-    let ingester2 = Ingester::with_shard_config(
-        IngesterConfig::default(),
-        object_store.clone(),
-        metadata.clone(),
-        storage_config_tenant2.clone(),
-        metric_schema,
-        Default::default(),
-        2, // tenant_id 2
-    );
-
-    // Subscribe to tenant 1 data
-    let mut rx_tenant1 = ingester1.subscribe_filtered(TopicFilter::Tenant(1)).await;
-
-    // Broadcast data from both tenants
-    let base_time = 1_000_000_000_000_000_000i64;
-
-    let batch1 = create_test_batch(vec![base_time], vec!["cpu"], vec![1.0], vec!["host-1"]);
-    let topic_batch1 = TopicBatch {
-        batch: batch1,
-        metadata: BatchMetadata {
-            shard_id: "shard-1".to_string(),
-            tenant_id: 1,
-            metrics: vec!["cpu".to_string()],
-        },
-    };
-
-    let batch2 = create_test_batch(
-        vec![base_time + 1],
-        vec!["memory"],
-        vec![2.0],
-        vec!["host-2"],
-    );
-    let topic_batch2 = TopicBatch {
-        batch: batch2,
-        metadata: BatchMetadata {
-            shard_id: "shard-2".to_string(),
-            tenant_id: 2,
-            metrics: vec!["memory".to_string()],
-        },
-    };
-
-    ingester1.topic_broadcast().send(topic_batch1).ok();
-    ingester2.topic_broadcast().send(topic_batch2).ok();
-
-    // Should receive only tenant 1 data
-    let result = tokio::time::timeout(Duration::from_millis(100), rx_tenant1.recv()).await;
-    assert!(result.is_ok(), "Should receive tenant 1 batch");
-
-    // Should not receive tenant 2 data
-    let result = tokio::time::timeout(Duration::from_millis(50), rx_tenant1.recv()).await;
-    assert!(result.is_err(), "Should not receive data from other tenant");
-
-    println!("✅ Tenant isolation verified");
-}
